requirements.txt
pip
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
oauth2client

